+++ src/services/ConversationRenderer.ts
@@ -1,4 +1,4 @@
-import { parseYaml, TFile } from 'obsidian';
+import { TFile } from 'obsidian';
 import { uniqueID } from '../utils/uniqueID';
 import { getTranslation } from '../i18n';
 import { ConversationMessage, ConversationRole } from '../types/types';
@@ -13,15 +13,20 @@ import { prependChunk } from 'src/utils/textStreamer';
 import type { ModelMessage, TextPart, FilePart, ImagePart, ReasoningOutput } from 'ai';
 import { ToolCallPart, ToolResultPart } from 'src/solutions/commands/tools/types';
 import { MarkdownUtil } from 'src/utils/markdownUtils';
-import { ArtifactType, ReadContentArtifactImpl } from 'src/solutions/artifact';
-import { removeUndefined } from 'src/utils/removeUndefined';
+import { ArtifactType } from 'src/solutions/artifact';
+import { applyMixins } from 'src/utils/applyMixins';
+import { ToolSerialization } from './ConversationRender/ToolSerialization';
+import { Frontmatter } from './ConversationRender/Frontmatter';
 import { Events } from 'src/types/events';
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface -- declaration merge: class body + mixin prototype
+export interface ConversationRenderer extends ToolSerialization, Frontmatter {}
 
 export class ConversationRenderer {
   static instance: ConversationRenderer;
   private streamingFiles = new Set<string>();
 
-  private constructor(private plugin: StewardPlugin) {}
+  private constructor(public readonly plugin: StewardPlugin) {}
 
   static getInstance(plugin?: StewardPlugin): ConversationRenderer {
     if (plugin) {
@@ -208,251 +213,6 @@ export class ConversationRenderer {
     const newLines = [...lines];
     newLines.splice(startLineIndex, endLineIndex - startLineIndex);
     return newLines.join('\n');
-  }
-
-  /**
-   * Serialize tool calls to a conversation note.
-   * The result could be inlined or referenced to a message or an artifact.
-   * @returns The message ID for referencing
-   */
-  public async serializeToolInvocation(params: {
-    path: string;
-    command?: string;
-    agent?: string;
-    text?: string;
-    handlerId?: string;
-    step?: number;
-    toolInvocations: (ToolCallPart | ToolResultPart)[];
-  }): Promise<string | undefined> {
-    try {
-      const file = this.getConversationFileByName(params.path);
-
-      // Get message metadata
-      const { messageId, comment } = await this.buildMessageMetadata(params.path, {
-        role: 'Assistant',
-        command: params.command,
-        agent: params.agent,
-        type: 'tool-invocation',
-        handlerId: params.handlerId,
-        step: params.step,
-      });
-
-      // Process the file content
-      await this.plugin.app.vault.process(file, currentContent => {
-        let contentToAdd = params.text ? `${params.text}\n` : '';
-
-        contentToAdd += `\`\`\`stw-artifact\n${JSON.stringify(params.toolInvocations)}\n\`\`\``;
-        return `${currentContent}\n\n${comment}\n${contentToAdd}`;
-      });
-
-      return messageId;
-    } catch (error: unknown) {
-      logger.error('Error serializing tool call:', error);
-      return undefined;
-    }
-  }
-
-  /**
-   * Deserialize tool calls from a message
-   */
-  public async deserializeToolInvocations(params: {
-    message: ConversationMessage;
-    conversationTitle: string;
-  }): Promise<Array<ToolCallPart | ToolResultPart | FilePart | ImagePart | TextPart> | null> {
-    // Extract all tool calls from the message content
-    const toolInvocationsMatches = params.message.content.match(
-      /```stw-artifact\n([\s\S]*?)\n```/g
-    );
-    if (!toolInvocationsMatches || toolInvocationsMatches.length === 0) {
-      logger.error('No stw-artifact blocks found in message content');
-      return null;
-    }
-
-    const toolInvocations: Array<ToolCallPart | ToolResultPart | FilePart | ImagePart | TextPart> =
-      [];
-
-    const resolveToolInvocation = async (
-      type: 'tool-call' | 'tool-result',
-      toolInvocation: ReturnType<typeof JSON.parse>
-    ): Promise<Array<ToolCallPart | ToolResultPart | FilePart | ImagePart | TextPart>> => {
-      switch (type) {
-        case 'tool-call': {
-          const input = toolInvocation.input ?? toolInvocation.args;
-          return [
-            {
-              type,
-              toolName: toolInvocation.toolName,
-              toolCallId: toolInvocation.toolCallId,
-              input,
-            },
-          ];
-        }
-
-        case 'tool-result':
-        default: {
-          // Resolve the result based on its type (Backward-compatible with AI SDK v4 format)
-          // let resolvedOutput: ToolResultPart['output'];
-          const output = toolInvocation.output ?? toolInvocation.result;
-          const additionalParts: Array<TextPart | FilePart | ImagePart> = [];
-
-          let resolvedOutput: ToolResultPart['output'] =
-            typeof output === 'string'
-              ? {
-                  type: 'text',
-                  value: output,
-                }
-              : output;
-
-          if (
-            resolvedOutput.type !== 'execution-denied' &&
-            typeof resolvedOutput.value === 'string'
-          ) {
-            // Check if it's an artifact reference
-            if (resolvedOutput.value.startsWith('artifactRef:')) {
-              const artifactId = resolvedOutput.value.substring('artifactRef:'.length);
-              const artifact = await this.plugin.artifactManagerV2
-                .withTitle(params.conversationTitle)
-                .getArtifactById(artifactId);
-              if (artifact) {
-                // If the artifact is marked as deleted, then skip sending data, return the deleteReason instead
-                if (artifact.deleteReason) {
-                  resolvedOutput = {
-                    type: 'json',
-                    value: removeUndefined({
-                      id: artifact.id,
-                      artifactType: artifact.artifactType,
-                      deleteReason: artifact.deleteReason,
-                      status: 'deleted', // Add this field
-                      // Don't send other fields of the artifact.
-                    }),
-                  };
-                } else {
-                  resolvedOutput = {
-                    type: 'json',
-                    value: removeUndefined(artifact),
-                  };
-
-                  // Check if artifact has imagePaths (from READ_CONTENT artifacts)
-                  // Include images in the same content as the tool-result
-                  if (
-                    artifact instanceof ReadContentArtifactImpl &&
-                    artifact.imagePaths &&
-                    artifact.imagePaths.length > 0
-                  ) {
-                    const imageParts = await this.plugin.userMessageService.getImagePartsFromPaths(
-                      artifact.imagePaths
-                    );
-                    for (const [path, imagePart] of imageParts) {
-                      additionalParts.push({ type: 'text', text: path }, imagePart);
-                    }
-                  }
-                }
-              } else {
-                logger.error(`Artifact not found: ${artifactId}`);
-                resolvedOutput = {
-                  type: 'error-text',
-                  value: `Artifact not found: ${artifactId}`,
-                };
-              }
-            } else if (resolvedOutput.value.startsWith('messageRef:')) {
-              const messageId = resolvedOutput.value.substring('messageRef:'.length);
-              const referencedMessage = await this.getMessageById(
-                params.conversationTitle,
-                messageId,
-                true // Exclude tool-hidden content when resolving message references
-              );
-              if (referencedMessage) {
-                resolvedOutput = {
-                  type: 'text',
-                  value: referencedMessage.content,
-                };
-              } else {
-                logger.error(`Message not found: ${messageId}`);
-                resolvedOutput = {
-                  type: 'error-text',
-                  value: `Message not found: ${messageId}`,
-                };
-              }
-            }
-            // Otherwise, it's an inlined result, keep as is
-          }
-
-          return [
-            {
-              type,
-              toolName: toolInvocation.toolName,
-              toolCallId: toolInvocation.toolCallId,
-              output: resolvedOutput,
-            },
-            ...additionalParts,
-          ];
-        }
-      }
-    };
-
-    for (const toolInvocationsMatch of toolInvocationsMatches) {
-      const toolInvocationData = JSON.parse(
-        toolInvocationsMatch.replace(/```stw-artifact\n|\n```/g, '')
-      );
-
-      // Ensure toolInvocationData is an array
-      if (!Array.isArray(toolInvocationData)) {
-        logger.error('Tool invocation data should be an array', toolInvocationData);
-        continue;
-      }
-
-      for (const toolInvocation of toolInvocationData) {
-        const { toolName, toolCallId, type } = toolInvocation;
-
-        if (!toolName || !toolCallId || !type) {
-          logger.error('Invalid tool call data structure', toolInvocation);
-          continue; // Skip invalid tool calls but continue processing others
-        }
-
-        if (type === 'tool-call') {
-          toolInvocations.push(...(await resolveToolInvocation('tool-call', toolInvocation)));
-          // In the old version, all types are tool-call, we need to handle tool-result here.
-          // Check for both output (new format) and result (old format) for backward compatibility
-          if (toolInvocation.output || toolInvocation.result) {
-            toolInvocations.push(...(await resolveToolInvocation('tool-result', toolInvocation)));
-          }
-        } else {
-          // In the new version, the input could be included, so handle tool-call here
-          if (toolInvocation.input) {
-            toolInvocations.push(...(await resolveToolInvocation('tool-call', toolInvocation)));
-          }
-          toolInvocations.push(...(await resolveToolInvocation('tool-result', toolInvocation)));
-        }
-      }
-    }
-
-    return toolInvocations.length > 0 ? toolInvocations : null;
-  }
-
-  /**
-   * Extracts tool names from a tool-invocation message content (sync, no artifact resolution).
-   * Used for filtering groups by compactability before full deserialization.
-   */
-  public extractToolNamesFromToolInvocation(content: string): string[] {
-    const matches = content.match(/```stw-artifact\n([\s\S]*?)\n```/g);
-    if (!matches?.length) return [];
-
-    const names: string[] = [];
-    for (const match of matches) {
-      try {
-        const jsonStr = match.replace(/```stw-artifact\n|\n```/g, '');
-        const data = JSON.parse(jsonStr);
-        if (!Array.isArray(data)) continue;
-        for (const item of data) {
-          if (item?.toolName && typeof item.toolName === 'string') {
-            names.push(item.toolName);
-          }
-        }
-      } catch {
-        // Skip malformed blocks
-      }
-    }
-    return names;
   }
 
   /**
@@ -1884,80 +1644,6 @@ export class ConversationRenderer {
   }
 
   /**
-   * Gets a property from the conversation's YAML frontmatter
-   * Tries cache first, then reads directly from file if not found
-   * @param conversationTitle The title of the conversation
-   * @param property The property name to retrieve
-   * @returns The property value or undefined if not found
-   */
-  public async getConversationProperty<T>(
-    conversationTitle: string,
-    property: string,
-    forceRefresh?: boolean
-  ): Promise<T | undefined> {
-    try {
-      const file = this.getConversationFileByName(conversationTitle);
-
-      // Try to get from cache first
-      const fileCache = this.plugin.app.metadataCache.getFileCache(file);
-
-      if (fileCache?.frontmatter && !forceRefresh) {
-        return fileCache.frontmatter[property];
-      }
-
-      // Cache miss, read directly from file
-      if (forceRefresh) {
-        logger.log(`Force refresh for property ${property}, reading directly from file`);
-      } else {
-        logger.log(`Cache miss for property ${property}, reading directly from file`);
-      }
-      const fileContent = await this.plugin.app.vault.read(file);
-      const frontmatterRegex = /^---\n([\s\S]*?)\n---/;
-      const match = fileContent.match(frontmatterRegex);
-
-      if (match) {
-        const frontmatter = parseYaml(match[1]) as Record<string, unknown>;
-        return frontmatter[property] as T;
-      }
-
-      return undefined;
-    } catch (error) {
-      logger.error(`Error getting conversation property ${property}:`, error);
-      return undefined;
-    }
-  }
-
-  /**
-   * Updates a property in the conversation's YAML frontmatter
-   * @param conversationTitle The title of the conversation
-   * @param properties The properties to update
-   * @returns True if successful, false otherwise
-   */
-  public async updateConversationFrontmatter(
-    conversationTitle: string,
-    properties: Array<{ name: string; value?: unknown; delete?: boolean }>
-  ): Promise<boolean> {
-    try {
-      const file = this.getConversationFileByName(conversationTitle);
-
-      await this.plugin.app.fileManager.processFrontMatter(file, frontmatter => {
-        for (const prop of properties) {
-          if (prop.delete) {
-            delete frontmatter[prop.name];
-            continue;
-          }
-          frontmatter[prop.name] = prop.value;
-        }
-      });
-
-      return true;
-    } catch (error) {
-      logger.error(`Error updating conversation frontmatter:`, error);
-      return false;
-    }
-  }
-
-  /**
    * Deletes a message and all messages below it from a conversation note
    * @param conversationTitle The title of the conversation
    * @param messageId The ID of the message to delete
@@ -2147,3 +1833,5 @@ export class ConversationRenderer {
     return file;
   }
 }
+
+applyMixins(ConversationRenderer, [ToolSerialization, Frontmatter]);
