+++ src/solutions/commands/agents/SuperAgent/SuperAgent.ts
@@ -21,6 +21,7 @@ import {
 } from '../agentTools';
 import type { AgentCorePromptContext } from '../../Agent';
 import { isGoogleModel } from '../googleUtils';
+import { USAGE_AGENT_KEY } from 'src/services/ConversationRender/Frontmatter';
 
 const SUPER_AGENT_VALID_TOOL_NAMES: ReadonlySet<ToolName> = SUPER_AGENT_TOOL_NAMES;
 
@@ -306,6 +307,16 @@ NOTE:
       toolCalls = result.toolCalls;
       conversationHistory = result.conversationHistory;
       toolContentStreamInfo = result.toolContentStreamInfo;
+      try {
+        await this.renderer.recordTokenUsage(
+          title,
+          USAGE_AGENT_KEY.super,
+          result.usage,
+          result.totalUsage
+        );
+      } catch (usageError) {
+        logger.error('Failed to record super agent token usage', usageError);
+      }
     }
 
     const toolProcessingResult = await this.executeToolCalls({
