+++ src/utils/applyMixins.ts
@@ -1,4 +1,8 @@
-import { Constructor } from 'obsidian';
+/**
+ * Class or constructor function with a `prototype` object.
+ * Structural type so classes with **private** constructors (e.g. singletons) type-check without casts.
+ */
+export type MixinConstructor = { prototype: object };
 
 /**
  * Apply mixins to a base class
@@ -6,10 +10,7 @@ import { Constructor } from 'obsidian';
  * @param derivedCtor - The derived constructor
  * @param constructors - Array of mixin constructors
  */
-export function applyMixins(
-  derivedCtor: Constructor<unknown>,
-  constructors: Constructor<unknown>[]
-): void {
+export function applyMixins(derivedCtor: MixinConstructor, constructors: MixinConstructor[]): void {
   for (const baseCtor of constructors) {
     for (const name of Object.getOwnPropertyNames(baseCtor.prototype)) {
       if (name === 'constructor') continue;
